move GetSubReshapeDimRanges to a single file
