+++ paddle/cinn/adt/get_sub_reshape_dim_ranges.h
@@ -1,0 +1,31 @@
+// Copyright (c) 2023 PaddlePaddle Authors. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+#include <optional>
+#include <vector>
+
+#include "paddle/cinn/adt/adt.h"
+
+namespace cinn::adt {
+
+class Constant;
+
+std::optional<std::tuple<std::vector<std::pair<int, int>>,
+                         std::vector<std::pair<int, int>>>>
+GetSubReshapeDimRanges(const List<Constant>& lhs_dims,
+                       const List<Constant>& rhs_dims);
+
+}  // namespace cinn::adt
